+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # fj-doc-guides changelog
 
+## 0.1.0-rc.002 - 2023-08-06
+* [Venus Guide E001 : Hello World PDF JSON Source](fj-doc-guides-E001-hello-world-pdf-json-source/README.md)
+* Started guidelines section
+
 ## 0.1.0-rc.001 - 2023-07-31
 * [Venus Guide A001 : Hello World PDF](fj-doc-guides-A001-hello-world-pdf/README.md)
 * Initial project setup